+++ qrust-domain/src/main/java/com/qrust/common/config/FeignConfig.java
@@ -1,6 +1,6 @@
-package com.qrust.external.google.feign.config;
+package com.qrust.common.config;
 
-import com.qrust.external.google.feign.client.GoogleSafeBrowsingFeignClient;
+import com.qrust.external.google.infrastructure.GoogleSafeBrowsingFeignClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
